key collision, API change, win check, fixed bug
Key collision added when player collides with key to destroy and pick up the key in essence.

API change in MazeRunnerMazeGeneration from getItemMap to getTileContentsMap with 2 parameters, begin and end, representing player and treasure chest in this case.

Checked win condition by only animating and giving a victory screen if player has all 4 keys.

Fixed bug where a treasure chest might take the spot of a key or player